requirements.txt
fastapi
uvicorn
python-multipart
aiohttp
numpy
jieba
rank-bm25
pymilvus
python-docx
pydantic
pydantic-settings
